requirements.txt
# Core ML/AI packages
ultralytics>=8.3.40
torch>=2.6.0
torchvision>=0.17.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0

# GUI and visualization
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Reporting
reportlab>=4.0.0

# Data and training
roboflow>=1.1.0 